Use exponential backoff for failed jobs.

+++ src/jobs/schema.sql
@@ -5,6 +5,6 @@ CREATE TABLE IF NOT EXISTS pg_queue_jobs (
     id BIGSERIAL NOT NULL,
     body bytea NOT NULL,
     retry_at TIMESTAMPTZ NOT NULL DEFAULT current_timestamp,
-    retried_count BIGINT NOT NULL DEFAULT 0,
+    retried_count INT NOT NULL DEFAULT 0,
     PRIMARY KEY (id)
 );
